refactor: move pkg tests to test
singular package names are better
