+++ controllers/routemonitor/supplement/supplement_test.go
@@ -25,7 +25,7 @@ import (
 	consterror "github.com/openshift/route-monitor-operator/pkg/const/test/error"
 	constinit "github.com/openshift/route-monitor-operator/pkg/const/test/init"
 	utilreconcile "github.com/openshift/route-monitor-operator/pkg/util/reconcile"
-	clientmocks "github.com/openshift/route-monitor-operator/pkg/util/tests/generated/mocks/client"
+	clientmocks "github.com/openshift/route-monitor-operator/pkg/util/test/generated/mocks/client"
 )
 
 var _ = Describe("Routemonitor", func() {
